+++ testing/UNKNOWN_ARGUMENT_Scenario.cpp
@@ -147,7 +147,7 @@ void build_UNKNOWN_ARGUMENT_scenario(Randomizer* rnd, ScenarioData& scenario){
             }
             break;    
         case APDataType::TEXT:
-            result_u32 = rnd->gen_integral_range<size_t>(1, MAX_TEXT_ARG_LEN);
+            result_u32 = rnd->gen_integral_range<uint32_t>(1, MAX_TEXT_ARG_LEN);
             value = rnd->gen_string(result_u32, nullptr);
             break;    
         case APDataType::FLAG:
